VPN changes for IPv6 ND based aliveness monitor
+ Learnt hidden IPv6 addresses are monitored using ND based aliveness
  monitor.
+ Tried to have common code for both IPv4 and IPv6 whereever possible.

Depends-On: https://git.opendaylight.org/gerrit/#/c/73005/

JIRA: NETVIRT-1213

+++ vpnmanager/impl/src/main/java/org/opendaylight/netvirt/vpnmanager/arp/responder/ArpResponderHandler.java
@@ -13,6 +13,7 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 import org.opendaylight.controller.md.sal.binding.api.DataBroker;
 import org.opendaylight.genius.interfacemanager.interfaces.IInterfaceManager;
+import org.opendaylight.genius.mdsalutil.NWUtil;
 import org.opendaylight.netvirt.elan.arp.responder.ArpResponderInput.ArpReponderInputBuilder;
 import org.opendaylight.netvirt.elan.arp.responder.ArpResponderUtil;
 import org.opendaylight.netvirt.elanmanager.api.IElanService;
@@ -131,7 +132,7 @@ public class ArpResponderHandler {
                 gatewayIp = gwIpOptional.get();
             }
         }
-        if (gatewayIp != null) {
+        if (gatewayIp != null && NWUtil.isIpv4Address(gatewayIp)) {
             ArpReponderInputBuilder builder = new ArpReponderInputBuilder();
             builder.setDpId(dpId).setInterfaceName(ifName).setSpa(gatewayIp).setLportTag(lportTag);
             elanService.removeArpResponderFlow(builder.buildForRemoveFlow());
